requirements.txt
langbot-plugin
pyyaml